+++ packages/markdown/markdown.module.ts
@@ -1,5 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
+
 import { NuMarkdownPreviewComponent } from './markdown-preview.component';
 import { NuMarkdownComponent } from './markdown.component';
 import { NuMarkdownConfig, NU_MARKDOWN_CONFIG } from './markdown.config';
@@ -8,7 +9,7 @@ const COMPONENTS = [NuMarkdownComponent, NuMarkdownPreviewComponent];
 
 @NgModule({
   imports: [CommonModule, ...COMPONENTS],
-  exports: COMPONENTS,
+  exports: COMPONENTS
 })
 export class NuMarkdownModule {
   /**
@@ -17,7 +18,7 @@ export class NuMarkdownModule {
   static forRoot(config?: NuMarkdownConfig): ModuleWithProviders<NuMarkdownModule> {
     return {
       ngModule: NuMarkdownModule,
-      providers: [{ provide: NU_MARKDOWN_CONFIG, useValue: config }],
+      providers: [{ provide: NU_MARKDOWN_CONFIG, useValue: config }]
     };
   }
 }