+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## <small>19.0.1 (2025-01-27)</small>
+
+* fix(monaco-editor): fix missing monaco types (#84) ([9c28df0](https://github.com/ng-util/ng-util/commit/9c28df0)), closes [#84](https://github.com/ng-util/ng-util/issues/84)
+
+
+
 ## 19.0.0 (2024-11-23)
 
 * build: bump to angular 19 (#81) ([7e6412d](https://github.com/ng-util/ng-util/commit/7e6412d)), closes [#81](https://github.com/ng-util/ng-util/issues/81)